fixed checking modified time in windows

+++ src/kmake.h
@@ -29,7 +29,7 @@ void install(struct Config* config);
 struct Config get_config(void);
 void free_config(struct Config config);
 const char* get_cwd(void);
-unsigned long get_file_mtime(const char* filename);
+unsigned long long get_file_mtime(const char* filename);
 
 str_array get_files(const char* dir);
 str_array get_source_files(const char* dir);
